Create files for game command
