+++ src/OpenLoco/src/GameCommands/GameCommands.cpp
@@ -61,6 +61,7 @@
 #include "Vehicles/VehiclePickup.h"
 #include "Vehicles/VehiclePickupAir.h"
 #include "Vehicles/VehiclePickupWater.h"
+#include "Vehicles/VehicleRearrange.h"
 #include "Vehicles/VehicleRefit.h"
 #include "Vehicles/VehicleReverse.h"
 #include "Vehicles/VehicleSell.h"
